+++ lib/datamappify/repository/query_method/update.rb
@@ -4,6 +4,10 @@ module Datamappify
   module Repository
     module QueryMethod
       class Update < Save
+        # @see Method#dirty_aware?
+        def dirty_aware?
+          true
+        end
       end
     end
   end